requirements.txt
apache-airflow-providers-snowflake==4.1.0
dbt-core==1.5.3
dbt-snowflake==1.5.3
snowflake-sqlalchemy==1.4.7
snowflake-connector-python==3.1.1